WIP correct type error and write tests

+++ rust_src/src/multibyte.rs
@@ -93,7 +93,9 @@ impl LispStringRef {
     /// considered; the validity of the following bytes is not checked.  Tabs in
     /// STRING are always taken to occupy `tab-width' columns.
     pub fn width(self) -> usize {
-        unsafe { lisp_string_width(self.as_lisp_obj(), -1, ptr::null_mut(), ptr::null_mut()) }
+        unsafe {
+            lisp_string_width(self.as_lisp_obj(), -1, ptr::null_mut(), ptr::null_mut()) as usize
+        }
     }
 
     pub fn is_multibyte(self) -> bool {
